share the parts context fetch with pathids prefetch
